going home

+++ libft/ft_strmap.c
@@ -6,7 +6,7 @@
 /*   By: mthoman <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/12 13:15:46 by mthoman           #+#    #+#             */
-/*   Updated: 2018/11/12 13:17:04 by mthoman          ###   ########.fr       */
+/*   Updated: 2018/11/12 16:45:14 by mthoman          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_strmap(char const *s, char (*f)(char))
 {
 	char	*fresh;
-	int		i;
+	size_t		i;
 
 	if (!(fresh = ft_strnew(ft_strlen(s))))
 		return (NULL);
